isolate bragg code into own source file
