+++ src/Grid.cpp
@@ -11,6 +11,7 @@
 #include "becke_partitioning.h"
 #include "error_handling.h"
 #include "grid_radial.h"
+#include "bragg.h"
 #include "lebedev/sphere_lebedev_rule.h"
 #include "parameters.h"
