+++ src/becke_partitioning.h
@@ -1,7 +1,3 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-
 #ifndef BECKE_PARTITIONING_H_INCLUDED
 #define BECKE_PARTITIONING_H_INCLUDED
 
@@ -24,6 +20,4 @@ double get_becke_w(const double x_coordinates_au[],
                    const double y,
                    const double z);
 
-double get_bragg_angstrom(const int charge);
-
-#endif
+#endif /* BECKE_PARTITIONING_H_INCLUDED */
