+++ src/bragg.h
@@ -1,0 +1,6 @@
+#ifndef BRAGG_H_INCLUDED
+#define BRAGG_H_INCLUDED
+
+double get_bragg_angstrom(const int charge);
+
+#endif /* BRAGG_H_INCLUDED */
